+++ modules/service/src/main/java/com/github/wezzen/insight/service/TagService.java
@@ -3,6 +3,8 @@ package com.github.wezzen.insight.service;
 import com.github.wezzen.insight.dto.response.TagDTO;
 import com.github.wezzen.insight.model.Tag;
 import com.github.wezzen.insight.repository.TagRepository;
+import com.github.wezzen.insight.service.exception.DuplicateTagException;
+import com.github.wezzen.insight.service.exception.TagNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,7 +25,7 @@ public class TagService {
     @Transactional
     public TagDTO createTag(final String tag) {
         if (tagRepository.findById(tag).isPresent()) {
-            throw new RuntimeException("Tag " + tag + " already exists");
+            throw new DuplicateTagException("Tag " + tag + " already exists");
         }
         final Tag saved = tagRepository.save(new Tag(tag));
         return new TagDTO(saved.getTag());
@@ -39,6 +41,9 @@ public class TagService {
 
     @Transactional
     public void deleteTag(final String tag) {
+        if (tagRepository.findById(tag).isEmpty()) {
+            throw new TagNotFoundException("Tag " + tag + " not found");
+        }
         tagRepository.deleteById(tag);
     }
 }
